refactor: Service Layer를 삭제하고 적절한 이름과 패키지 경로를 갖도록 변경
