+++ src/main/java/RaceApplicationConfig.java
@@ -3,7 +3,7 @@ import controller.RaceController;
 import controller.RaceControllerHandler;
 import java.lang.reflect.Proxy;
 import java.util.Scanner;
-import service.RaceService;
+import model.RaceGame;
 import utils.RacingRandomNumberGenerator;
 import view.InputView;
 import view.OutputView;
@@ -12,10 +12,10 @@ public class RaceApplicationConfig {
 
     public RaceController raceController() {
         RacingRandomNumberGenerator generator = new RacingRandomNumberGenerator();
-        RaceService raceService = new RaceService(generator);
+        RaceGame raceGame = new RaceGame(generator);
         InputView inputView = new InputView(new Scanner(System.in));
         OutputView outputView = new OutputView();
-        RaceController raceController = new NumberRaceController(inputView, outputView, raceService);
+        RaceController raceController = new NumberRaceController(inputView, outputView, raceGame);
 
         return (RaceController) Proxy.newProxyInstance(RaceController.class.getClassLoader(),
                 new Class[]{RaceController.class},
